# shorty-server/docs/swagger.yaml
definitions:
  models.Redirect:
    properties:
      path:
        description: key/id
        type: string
      url:
        type: string
    type: object
  models.RedirectUser:
    properties:
      email:
        type: string
    type: object
  models.Response:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    name: Containerplattformen
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /{path}:
    get:
      consumes:
      - text/html
      description: redirects to the URL
      parameters:
      - description: Path
        in: path
        name: path
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: Found
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Redirect
      tags:
      - redirect
  /admin/:
    get:
      consumes:
      - application/json
      description: gets all redirects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Redirect'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
      summary: Get redirect
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: adds a redirect to url
      parameters:
      - description: Query
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/models.Redirect'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
      summary: Add redirect
      tags:
      - admin
  /admin/{id}:
    delete:
      consumes:
      - application/json
      description: delets a redirect by id
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
      summary: Delete redirect
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: Updates a redirect to given url
      parameters:
      - description: Query
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/models.Redirect'
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
      summary: Updates redirect
      tags:
      - admin
  /admin/qr/{id}:
    get:
      consumes:
      - application/json
      description: gets qrcode by id
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
      summary: Get qr-code by id
      tags:
      - admin
  /admin/user:
    get:
      consumes:
      - application/json
      description: Returns a list of all redirect entries configured for the admin
        panel.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RedirectUser'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
      summary: Get all admin
      tags:
      - admin user
    post:
      consumes:
      - application/json
      description: adds a admin user
      parameters:
      - description: Query
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/models.RedirectUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RedirectUser'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
      summary: Add admin user
      tags:
      - admin user
  /admin/user/{id}:
    delete:
      consumes:
      - application/json
      description: delets a admin user by email
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RedirectUser'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
      summary: Delete admin user
      tags:
      - admin user
  /qr/:
    get:
      consumes:
      - application/json
      description: get qrcode by query
      parameters:
      - description: Query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get qrcode by query
      tags:
      - qr
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"